Continue with renaming including the paths
