+++ jdaviz/configs/cubeviz/cubeviz.yaml
@@ -21,7 +21,7 @@ tray:
   - g-metadata-viewer
   - g-plot-options
   - g-data-quality
-  - g-subset-plugin
+  - g-subset-tools
   - g-markers
   - cube-slice
   - g-unit-conversion
