+++ jdaviz/configs/default/plugins/__init__.py
@@ -3,7 +3,7 @@ from .gaussian_smooth.gaussian_smooth import *  # noqa
 from .data_tools import *  # noqa
 from .viewer_creator import *  # noqa
 from .subset_tools.subset_tools import *  # noqa
-from .subset_plugin.subset_plugin import * # noqa
+from .subset_widget.subset_widget import * # noqa
 from .model_fitting.model_fitting import *  # noqa
 from .collapse.collapse import *  # noqa
 from .line_lists.line_lists import *  # noqa
