+++ jdaviz/configs/default/plugins/subset_widget/subset_widget.vue
@@ -1,0 +1,7 @@
+<template>
+    <v-toolbar-items>
+        <g-subset-select></g-subset-select>
+        <g-subset-mode></g-subset-mode>
+        <!-- <v-divider vertical></v-divider> -->
+    </v-toolbar-items>
+</template>
