+++ jdaviz/configs/imviz/imviz.yaml
@@ -23,7 +23,7 @@ tray:
   - imviz-orientation
   - g-plot-options
   - g-data-quality
-  - g-subset-plugin
+  - g-subset-tools
   - g-markers
   - imviz-compass
   - imviz-line-profile-xy
