+++ jdaviz/configs/rampviz/rampviz.yaml
@@ -21,7 +21,7 @@ tray:
   - g-metadata-viewer
   - g-plot-options
   - g-data-quality
-  - g-subset-plugin
+  - g-subset-tools
   - ramp-extraction
   - g-markers
   - cube-slice
