+++ jdaviz/configs/specviz/specviz.yaml
@@ -19,7 +19,7 @@ toolbar:
 tray:
   - g-metadata-viewer
   - g-plot-options
-  - g-subset-plugin
+  - g-subset-tools
   - g-markers
   - g-gaussian-smooth
   - g-model-fitting
