+++ jdaviz/configs/specviz2d/specviz2d.yaml
@@ -17,7 +17,7 @@ toolbar:
 tray:
   - g-metadata-viewer
   - g-plot-options
-  - g-subset-plugin
+  - g-subset-tools
   - g-markers
   - spectral-extraction
   - g-gaussian-smooth
